refactor: remove xontrib_meta module
this is no longer needed and can be merged with xontribs
